FPLA-2948: merge duplication of code in common component
