+++ service/src/main/java/uk/gov/hmcts/reform/fpl/enums/orders/ManageOrdersEndDateType.java
@@ -1,7 +1,8 @@
 package uk.gov.hmcts.reform.fpl.enums.orders;
 
-public enum ManageOrdersEndDateTypeWithEndOfProceedings {
+public enum ManageOrdersEndDateType {
     CALENDAR_DAY,
     CALENDAR_DAY_AND_TIME,
+    NUMBER_OF_MONTHS,
     END_OF_PROCEEDINGS
 }
