--- service/src/main/java/uk/gov/hmcts/reform/fpl/enums/orders/ManageOrdersEndDateTypeWithMonth.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.fpl.enums.orders;
-
-public enum ManageOrdersEndDateTypeWithMonth {
-    CALENDAR_DAY,
-    CALENDAR_DAY_AND_TIME,
-    NUMBER_OF_MONTHS
-}
