+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/event/ManageOrdersEventData.java
@@ -4,8 +4,7 @@ import lombok.Builder;
 import lombok.Value;
 import lombok.extern.jackson.Jacksonized;
 import uk.gov.hmcts.reform.fpl.enums.EPOType;
-import uk.gov.hmcts.reform.fpl.enums.orders.ManageOrdersEndDateTypeWithEndOfProceedings;
-import uk.gov.hmcts.reform.fpl.enums.orders.ManageOrdersEndDateTypeWithMonth;
+import uk.gov.hmcts.reform.fpl.enums.orders.ManageOrdersEndDateType;
 import uk.gov.hmcts.reform.fpl.model.Address;
 import uk.gov.hmcts.reform.fpl.model.common.DocumentReference;
 import uk.gov.hmcts.reform.fpl.model.common.dynamic.DynamicList;
@@ -35,8 +34,8 @@ public class ManageOrdersEventData {
     DocumentReference manageOrdersPowerOfArrest;
     Address manageOrdersEpoRemovalAddress;
     LocalDate manageOrdersExclusionStartDate;
-    ManageOrdersEndDateTypeWithMonth manageOrdersEndDateTypeWithMonth;
-    ManageOrdersEndDateTypeWithEndOfProceedings manageOrdersEndDateTypeWithEndOfProceedings;
+    ManageOrdersEndDateType manageOrdersEndDateTypeWithMonth;
+    ManageOrdersEndDateType manageOrdersEndDateTypeWithEndOfProceedings;
     LocalDate manageOrdersSetDateEndDate;
     LocalDateTime manageOrdersSetDateAndTimeEndDate;
     Integer manageOrdersSetMonthsEndDate;
